Split each Ui into seperate file
