--- ui_elements/passwordbuttons.py
@@ -1,185 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file 'passwordbuttons.ui'
-#
-# Created by: PyQt5 UI code generator 5.10
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_PasswordWindow(object):
-    def setupUi(self, PasswordWindow):
-        PasswordWindow.setObjectName("PasswordWindow")
-        PasswordWindow.resize(400, 316)
-        PasswordWindow.setMinimumSize(QtCore.QSize(400, 316))
-        PasswordWindow.setMaximumSize(QtCore.QSize(400, 316))
-        PasswordWindow.setStyleSheet("QWidget\n"
-"{\n"
-"    color: rgb(0, 123, 255);    \n"
-"    background-color: rgb(0, 0, 0);\n"
-"\n"
-"}\n"
-"\n"
-"QPushButton {\n"
-"    background-color: rgb(0, 0, 0);\n"
-"    color: (0, 123, 255);\n"
-"    border-width: 4px;\n"
-"    border-color: rgb(0, 123, 255);\n"
-"    border-style: solid;\n"
-"    border-radius: 7;\n"
-"    padding: 3px;\n"
-"    padding-left: 5px;\n"
-"    padding-right: 5px;\n"
-"}\n"
-"\n"
-"QPushButton:pressed\n"
-"{\n"
-"    color: rgb(239, 151, 0);    \n"
-"    border-color: rgb(239, 151, 0);\n"
-"}")
-        self.centralwidget = QtWidgets.QWidget(PasswordWindow)
-        self.centralwidget.setObjectName("centralwidget")
-        self.gridLayout = QtWidgets.QGridLayout(self.centralwidget)
-        self.gridLayout.setObjectName("gridLayout")
-        self.PB2 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB2.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB2.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB2.setFont(font)
-        self.PB2.setObjectName("PB2")
-        self.gridLayout.addWidget(self.PB2, 0, 1, 1, 1)
-        self.PB4 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB4.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB4.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB4.setFont(font)
-        self.PB4.setObjectName("PB4")
-        self.gridLayout.addWidget(self.PB4, 1, 0, 1, 1)
-        self.PB6 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB6.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB6.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB6.setFont(font)
-        self.PB6.setObjectName("PB6")
-        self.gridLayout.addWidget(self.PB6, 1, 2, 1, 1)
-        self.PB9 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB9.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB9.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB9.setFont(font)
-        self.PB9.setObjectName("PB9")
-        self.gridLayout.addWidget(self.PB9, 2, 2, 1, 1)
-        self.PB5 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB5.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB5.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB5.setFont(font)
-        self.PB5.setObjectName("PB5")
-        self.gridLayout.addWidget(self.PB5, 1, 1, 1, 1)
-        self.PB7 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB7.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB7.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB7.setFont(font)
-        self.PB7.setObjectName("PB7")
-        self.gridLayout.addWidget(self.PB7, 2, 0, 1, 1)
-        self.PB8 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB8.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB8.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB8.setFont(font)
-        self.PB8.setObjectName("PB8")
-        self.gridLayout.addWidget(self.PB8, 2, 1, 1, 1)
-        self.PB3 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB3.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB3.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB3.setFont(font)
-        self.PB3.setObjectName("PB3")
-        self.gridLayout.addWidget(self.PB3, 0, 2, 1, 1)
-        self.PB1 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB1.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB1.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB1.setFont(font)
-        self.PB1.setObjectName("PB1")
-        self.gridLayout.addWidget(self.PB1, 0, 0, 1, 1)
-        self.PBenter = QtWidgets.QPushButton(self.centralwidget)
-        self.PBenter.setMinimumSize(QtCore.QSize(125, 70))
-        self.PBenter.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(28)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PBenter.setFont(font)
-        self.PBenter.setObjectName("PBenter")
-        self.gridLayout.addWidget(self.PBenter, 3, 0, 1, 1)
-        self.PB0 = QtWidgets.QPushButton(self.centralwidget)
-        self.PB0.setMinimumSize(QtCore.QSize(125, 70))
-        self.PB0.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(40)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PB0.setFont(font)
-        self.PB0.setObjectName("PB0")
-        self.gridLayout.addWidget(self.PB0, 3, 1, 1, 1)
-        self.PBdel = QtWidgets.QPushButton(self.centralwidget)
-        self.PBdel.setMinimumSize(QtCore.QSize(125, 70))
-        self.PBdel.setMaximumSize(QtCore.QSize(125, 70))
-        font = QtGui.QFont()
-        font.setPointSize(28)
-        font.setBold(True)
-        font.setWeight(75)
-        self.PBdel.setFont(font)
-        self.PBdel.setObjectName("PBdel")
-        self.gridLayout.addWidget(self.PBdel, 3, 2, 1, 1)
-        PasswordWindow.setCentralWidget(self.centralwidget)
-
-        self.retranslateUi(PasswordWindow)
-        QtCore.QMetaObject.connectSlotsByName(PasswordWindow)
-
-    def retranslateUi(self, PasswordWindow):
-        _translate = QtCore.QCoreApplication.translate
-        PasswordWindow.setWindowTitle(_translate("PasswordWindow", "Passworteingabe"))
-        self.PB2.setText(_translate("PasswordWindow", "2"))
-        self.PB4.setText(_translate("PasswordWindow", "4"))
-        self.PB6.setText(_translate("PasswordWindow", "6"))
-        self.PB9.setText(_translate("PasswordWindow", "9"))
-        self.PB5.setText(_translate("PasswordWindow", "5"))
-        self.PB7.setText(_translate("PasswordWindow", "7"))
-        self.PB8.setText(_translate("PasswordWindow", "8"))
-        self.PB3.setText(_translate("PasswordWindow", "3"))
-        self.PB1.setText(_translate("PasswordWindow", "1"))
-        self.PBenter.setText(_translate("PasswordWindow", "Enter"))
-        self.PB0.setText(_translate("PasswordWindow", "0"))
-        self.PBdel.setText(_translate("PasswordWindow", "del"))
-
